+++ VComponents/Components/Lists/VSectionList/VSectionListSection.swift
@@ -10,6 +10,6 @@ import Foundation
 // MARK:- V Section List Section ViewModelable
 /// ViewModel that creates `VSectionList` sections
 public protocol VSectionListSectionViewModelable: Identifiable {
-    associatedtype VSectionListRow: VComponents.VSectionListRowViewModelable
-    var rows: [VSectionListRow] { get }
+    associatedtype VSectionListRowViewModelable: Identifiable
+    var rows: [VSectionListRowViewModelable] { get }
 }